longer

+++ src/Main.java
@@ -7,7 +7,7 @@ public class Main {
 
         System.out.println("Введите размерность массива продаж:");
         int size = scanner.nextInt();
-        int sales[]=new int[size];
+        long sales[]=new long[size];
 
         System.out.println("Введите суммы продаж, для завершения введите:");
         for (int i = 0; i < size; i++) {
@@ -16,7 +16,9 @@ public class Main {
         }
 
         SalesManager sale = new SalesManager( sales);
-        int max=sale.max();
+
+        long max=sale.max();
+
         System.out.println("Сумма максимальной продажи  "+max);
     }
 }
